requirements.txt
PyYAML==3.11
cffi==1.2.1
click==5.1
click-plugins==1.0.1
coverage==3.7.1
cryptography==1.0.1
enum34==1.0.4
idna==2.0
ipaddress==1.0.14
ndg-httpsclient==0.4.0
nose==1.3.7
psycopg2==2.6.1
pyOpenSSL==0.15.1
pyasn1==0.1.8
pycparser==2.14
python-dateutil==2.4.2
pytz==2015.4
requests==2.7.0
six==1.9.0
slacker==0.7.3
https://github.com/koordinates/python-client/tarball/master#egg=koordinates
